Remove useless page
