+++ app/page.tsx
@@ -1,12 +1,17 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
+import { Container, Stack } from '@mantine/core';
+import { Layout } from '@/components/Layout/Layout';
+import ProposalList from '@/components/Proposals/ProposalList';
 
-export default function HomePage() {
-  const router = useRouter();
-  useEffect(() => {
-    router.push('/proposals');
-  }, []);
-  return <></>;
+export default function ProposalsPage() {
+  return (
+    <Layout>
+      <Container p="0">
+        <Stack gap="15">
+          <ProposalList />
+        </Stack>
+      </Container>
+    </Layout>
+  );
 }
